Create verify-email.php
When the user clicks the verification link sent to their email inbox. It will direct them to this page which will compare the md5 HASH to the md5 users email HASH stored in the database. If the HASH's match, it will change the verification value for that email in the database